requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
python-dotenv==1.2.1
supabase==2.28.0
sentence-transformers==5.2.3
jellyfish==1.2.1
rapidfuzz==3.14.3
Unidecode==1.4.0
langdetect==1.0.9
deep-translator==1.11.4
